bpf-map-memory.py: Script for checking BPF maps memory usage
Script which returns the percentage of BPF maps memory usage (as an
integer). It it's about to reach 100, bpf-map-dynamic-size-ratio (or
bpf-policy-map-max, if dynamic size ratio is not used) has to be
increased.

The way this script works depends on Cilium configuration.

If bpf-map-dynamic-size-ratio is used, this script checks limits for all
dynamically allocated BPF maps (conntrack, nat, poicy) based on that
ratio, then checks the usage percentage for each of them and returns the
highest one.

If bpf-map-dynamic-size ratio is not used, this script checks the limit
set by bpf-policy-map-max, checks the numbber of elements in each policy
map and returns the percentage based on the highest number of elements
(so it returns the percentage of the most heavily used BPF map).

In both cases, that percentage is counted only on the single node. This
script is supposed to be used on all nodes (both control plane and
workers).

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>